task6-KAN-37 rebase from task 5
