+++ src/components/confirm-modal/reset-modal.css
