+++ src/components/modal/Modal.tsx
@@ -1,104 +1,25 @@
-import { FC, useEffect, useRef } from 'react';
-import './reset-modal.css';
+import React from 'react';
+import './modal.css';
+import RemoveIcon from "../../../public/assets/close.svg";
 
 interface ModalProps {
   show: boolean;
   title: string;
-  message: string;
-  onConfirm: () => void;
-  onCancel: () => void;
-  confirmLabel?: string;
-  cancelLabel?: string;
+  children: React.ReactNode;
+  onClose: () => void;
 }
 
-const Modal: FC<ModalProps> = ({
-  show,
-  title,
-  message,
-  onConfirm,
-  onCancel,
-  confirmLabel = 'Confirm',
-  cancelLabel = 'Cancel'
-}) => {
-  const modalRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const handleEscape = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') {
-        onCancel();
-      }
-    };
-
-    const handleClickOutside = (event: MouseEvent) => {
-      if (modalRef.current && !modalRef.current.contains(event.target as Node)) {
-        onCancel();
-      }
-    };
-
-    if (show) {
-      document.addEventListener('keydown', handleEscape);
-      document.addEventListener('mousedown', handleClickOutside);
-      document.body.style.overflow = 'hidden';
-    } else {
-      document.removeEventListener('keydown', handleEscape);
-      document.removeEventListener('mousedown', handleClickOutside);
-      document.body.style.overflow = 'unset';
-    }
-
-    return () => {
-      document.removeEventListener('keydown', handleEscape);
-      document.removeEventListener('mousedown', handleClickOutside);
-      document.body.style.overflow = 'unset';
-    };
-  }, [show, onCancel]);
-
-  useEffect(() => {
-    if (show && modalRef.current) {
-      const focusableElements = modalRef.current.querySelectorAll(
-        'button, [href], input, select, textarea, [tabindex]:not([tabindex="-1"])'
-      ) as NodeListOf<HTMLElement>;
-      const firstElement = focusableElements[0];
-      const lastElement = focusableElements[focusableElements.length - 1];
-
-      const trapFocus = (event: KeyboardEvent) => {
-        if (event.key !== 'Tab') return;
-        if (event.shiftKey) {
-          if (document.activeElement === firstElement) {
-            lastElement.focus();
-            event.preventDefault();
-          }
-        } else {
-          if (document.activeElement === lastElement) {
-            firstElement.focus();
-            event.preventDefault();
-          }
-        }
-      };
-
-      document.addEventListener('keydown', trapFocus);
-      firstElement.focus();
-
-      return () => {
-        document.removeEventListener('keydown', trapFocus);
-      };
-    }
-  }, [show]);
-
+const Modal: React.FC<ModalProps> = ({ show, title, children, onClose }) => {
   if (!show) return null;
 
   return (
-    <div className="modal-overlay" aria-modal="true">
-      <div className="modal" ref={modalRef}>
-        <h2>{title}</h2>
-        <p>{message}</p>
-        <div className="modal-actions">
-          <button className="modal-button confirm" onClick={onConfirm}>
-            {confirmLabel}
-          </button>
-          <button className="modal-button cancel" onClick={onCancel}>
-            {cancelLabel}
-          </button>
+    <div className="modal-backdrop">
+      <div className="modal">
+        <div className="modal-header">
+          <h2>{title}</h2>
+          <button onClick={onClose}><RemoveIcon/></button>
         </div>
+        <div className="modal-content">{children}</div>
       </div>
     </div>
   );
